WEBNEW-180 :white_check_mark: Added test for click handler
